Partially use node-jslint (fix #1)
The use of napa to install the jslint dependence was bringing installation problems, on visibly all platforms (see #1).
By using node-jslint (simply "jslint" on NPM), we get around the problematic command.
The possible next (big) step is to actually use the node-jslint package, instead of just using one of the files it provides.