Get rid of the minified version
In the last tag, the main file has a new feature (see #5), but the minified file is not up-to-date.
If I'm not mistaken, this brings this problem : since the "main" file in package.json is the minified version, the new feature is not available when you use this plugin.

My suggestion is to simply delete it. Thus this kind of problem will not be possible anymore.

Since this minified file is produced by the default gulp.task, and we don't want it, it has been added to .gitignore. This kinda make the gulpfile useless, so it will probably be good to delete or refactor it when the real tests are published.

Don't forget to add 0.0.7 tag if you merge this request.